一部処理群を外部モジュール化
